move base service to temp main app utils.py
